requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.47.0
django-filter==21.1
django-widgets-improved==1.5.0
djangorestframework==3.12.4
ecdsa==0.17.0
idna==3.3
oauthlib==3.1.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.2
ulid-py==1.1.0
urllib3==1.26.7
